+++ src/main/java/com/carret/market/web/member/PointController.java
@@ -3,8 +3,8 @@ package com.carret.market.web.member;
 import com.carret.market.application.member.MemberService;
 import com.carret.market.support.authorization.AuthenticationPrincipal;
 import com.carret.market.support.user.UserDetail;
-import com.carret.market.web.member.dto.PointRequest;
-import com.carret.market.web.member.dto.PointResponse;
+import com.carret.market.application.member.dto.PointRequest;
+import com.carret.market.application.member.dto.PointResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -17,23 +17,26 @@ import org.springframework.web.bind.annotation.RequestBody;
 @RequiredArgsConstructor
 public class PointController {
 
+    private static final String POINT_SUCCESS_MESSAGE = "정상적으로 포인트가 적립되었습니다.";
+
     private final MemberService memberService;
 
     @GetMapping("/point")
-    public String pointForm(@AuthenticationPrincipal UserDetail userDetail, Model model) {
+    public String pointForm(@AuthenticationPrincipal UserDetail userDetail,
+                            Model model) {
 
-        model.addAttribute("memberPointInfo", memberService.selectPoint(userDetail.getMemberDetail().getId()));
+        model.addAttribute("memberPointInfo", memberService.findPoint(userDetail.getMemberDetail().getId()));
 
         return "member/pointPage";
     }
 
     @PostMapping("/point")
     public ResponseEntity<PointResponse> pointCharge(@RequestBody PointRequest pointRequest,
-        @AuthenticationPrincipal UserDetail userDetail) {
+                                                    @AuthenticationPrincipal UserDetail userDetail) {
 
         memberService.pointCharge(pointRequest, userDetail.getMemberDetail().getId());
 
-        return ResponseEntity.ok(new PointResponse(true, "정상적으로 포인트가 적립되었습니다."));
+        return ResponseEntity.ok(new PointResponse(true, POINT_SUCCESS_MESSAGE));
     }
 
 }
